requirements.txt
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
datasets>=2.14.0
wikipedia>=1.4.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0

